Base Class for argoFloats Objects

In the normal situation, users will not create these objects directly. Instead, they are created by functions such as getIndex().

Slots

data

The data slot is a list with contents that vary with the object type. For example, getIndex() creates objects of type "index" that have a single unnamed element in data that is a data frame. This data frame has a column named file that is used in combination with metadata@ftpRoot to form a URL for downloading, along with columns named date, latitude, longitude, ocean, profiler_type, institution and date_update. Other "get" functions create objects with different contents.

metadata

The metadata slot is a list containing information about the data. The contents vary between objects and object types. That type is indicated by elements named type and subtype, which are checked by many functions within the package.

processingLog

The processingLog slot is a list containing time-stamped processing steps that may be stored in the object by argoFloats functions. These are noted in summary() calls.

Examples

str(new("argoFloats"))

#> Formal class 'argoFloats' [package "argoFloats"] with 3 slots
#>   ..@ metadata     :List of 2
#>   .. ..$ type   : chr "unspecified"
#>   .. ..$ subtype: chr "cycles"
#>   ..@ data         : list()
#>   ..@ processingLog:List of 2
#>   .. ..$ time : POSIXct[1:1], format: "2021-09-07 19:39:26"
#>   .. ..$ value: chr "create 'argoFloats' object"