Possibly print debugging information.

This function is intended mainly for use within the package, but users may also call it directly in their own code. Within the package, the value of debug is generally reduced by 1 on each nested function call, leading to indented messages. Most functions start and end with a call to argoFloatsDebug() that has style="bold" and unindent=1.

argoFloatsDebug(
  debug = 0,
  ...,
  style = "plain",
  showTime = FALSE,
  unindent = 0
)

Arguments

debug
an integer specifying the level of debugging. Values greater than zero indicate that some printing should be done. Values greater than 3 are trimmed to 3. Many functions pass debug=debug-1 down to deeper functions, which yields a nesting-indent format in the output.

...
values to be printed, analogous to the ... argument list of cat().

style
character value indicating special formatting, with "plain" for normal text, "bold" for bold-faced text, "italic" for italicized text, "red" for red text, "green" for green text, or "blue" for blue text. These codes may not be combined.

showTime
logical value indicating whether to preface message with the present time. This can be useful for learning about which operations are using the most time, but the default is not to show this, in the interests of brevity.

unindent
integer specifying the degree of reverse indentation to be done, as explained in the “Details” section.

Author

Dan Kelley

Examples

if (FALSE) {
argoFloatsDebug(1, "plain text\n")
argoFloatsDebug(1, "red text\n", style="red")
argoFloatsDebug(1, "blue text\n", style="blue")
argoFloatsDebug(1, "bold text\n", style="bold")
argoFloatsDebug(1, "italic text with time stamp\n", style="italic", showTime=TRUE)
}