Get Data for an Argo Float Profile

Get Data for an Argo Float Profile

getProfileFromUrl(
  url = NULL,
  destdir = "~/data/argo",
  destfile = NULL,
  force = FALSE,
  retries = 3,
  quiet = FALSE,
  debug = 0
)

Arguments

url
character value giving the URL for an Argo float a profile.

destdir
character value indicating the directory in which to store downloaded files. The default value of "." means to store the downloaded file in the present working directory. Set destdir=NULL if destfile is a filename with full path information. File clutter is reduced by creating a top-level directory called data, with subdirectories for various file types; see “Examples”.

destfile
optional character value that specifies the name to be used for the downloaded file. If this is not specified, then a name is determined from the value of url.

force
a logical value that indicates whether to force a file download, even if the destination file already exists.

retries
integer telling how many times to retry a download, if the first attempt fails.

quiet
logical value, passed to curl::curl_download(), that indicates whether to download silently. Since some downloads can be slow, the default value is quiet=FALSE.

debug
integer value indicating level of debugging. If this is less than 1, no debugging is done. Otherwise, some functions will print debugging information. If a function call fails, the first step should be to rerun the function with debug=1, to see if the output suggests a problem in the call.

Examples

if (FALSE) {
# These examples assume that the ~/data/argo directory exists.
library(argoFloats)
library(oce)

# Example 1: a particular file
url <- "ftp://ftp.ifremer.fr/ifremer/argo/dac/nmdis/2901633/profiles/R2901633_071.nc"
file <- getProfileFromUrl(url=url, destdir="~/data/argo")
argo <- read.argo(file)
plot(argo, which=c(1, 4, 6, 5))

# Example 2: argo profile nearest Sable Island
index <- getIndex(destdir="~/data/argo")
lon0 <- -59.9149
lat0 <- 43.9337
dist <- geodDist(index[["longitude"]], index[["latitude"]], lon0, lat0)
w <- which.min(dist)
url <- paste0(index[["metadata"]]["ftpRoot"], "/", index[["file"]][w])
fileSable <- getProfileFromUrl(url=url, destdir="~/data/argo")
argoSable <- read.oce(fileSable)
plot(argoSable, which=c(1, 4, 6, 5))
}