Convert Hexadecimal Digit to Integer Vector

hexToBits converts a string holding hexadecimal digits to a sequence of integers 0 or 1, for the bits. This is mainly for for use within showQCTests().

hexToBits(hex)

Arguments

hex
a vector of character values corresponding to a sequence of one or more hexadecimal digits (i.e. "0" through "9","a" through "f", or "A" through "F").

Value

An integer vector holding the bits as values 0 or 1. The inverse of 'mathematical' order is used, as is the case for the base R function rawToBits(); see the “Examples”.

Author

Jaimie Harbin and Dan Kelley

Examples

library(argoFloats)
hexToBits('3')    # 1 1 0 0

#> [1] 1 1 0 0

hexToBits('4000') # 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0

#>  [1] 0 0 0 0 0 0 0 0 0 0 0 0 0 0 1 0