Plot a TS diagram for an argoFloats object

Plot a TS diagram for an argoFloats object that was created with readProfiles(). This function is called by plot,argoFloats-method(), but may also be called directly.

plotArgoTS(
  x,
  xlim = NULL,
  ylim = NULL,
  type = "p",
  cex = 1,
  col = NULL,
  pch = 1,
  bg = "white",
  eos = "gsw",
  TSControl = NULL,
  debug = 0,
  ...
)

Arguments

x

an argoFloats object that was created with readProfiles().

xlim, ylim

optional limits of salinity and temperature axes; if not provided, the plot region will be large enough to show all the data.

type

character value indicating the of plot. This has the same meaning as for general R plots: "p" for points, "l" for lines, etc. Note that lines are joined up between cycles, unless the TSControl parameter indicates otherwise.

cex, col, pch, bg

values that have the same meaning as for general R plots if TSControl$groupByCycle is FALSE, except that if col is not provided by the user and if type is "l", then the points are colour-coded to indicate the value of data-quality flags. Black symbols indicate good data, i.e. data for which the flags for pressure, salinity and temperature are all in the set (1, 2, 5, 8). Red is used for bad data, with any of these three variables being flagged in the set (3, 4, 6, 7). And, finally gray is used if data that have not been assessed for quality, with the flags for all three of these variables being 0. (For more on flags, see Reference Table 2 in Section 3.2 of reference 1.)

eos

character value, either "gsw" (the default) for the Gibbs-Seawater (TEOS-10) equation of state or "unesco" for the 1980s-era UNESCO equation of state.

TSControl

an optional list that may have a logical element named groupByCycle, meaning to group the data by cycle. If TSControl is not provided, it is set to list(groupByDefault=FALSE). In grouped cases, the values of cex, col, and pch are passed to rep() to achieve the same length as the number of cycles in x. This can be useful in distinguishing between cycles.

debug

an integer controlling how much information is to be printed during operation. Use 0 for silent work, 1 for some information and 2 for more information. Note that plot,argoFloats-method() reduces its debug value by 1 before passing to plotArgoTS().

...

extra arguments provided to oce::plotTS().

References

Argo Data Management. “Argo User’s Manual.” Ifremer, July 5, 2022. https://doi.org/10.13155/29825.

Author

Dan Kelley and Jaimie Harbin