Look up a Value Within an argoFloats Object

This function provides an easy way to look up values within an argoFloats object, without the need to know the exact structure of the data. The action taken by [[ depends on the type element in the metadata slot of the object (which is set by the function that created the object), on the value of i and, in some cases, on the value of j; see “Details”.

# S4 method for argoFloats
[[(x, i, j, ...)

Arguments

x
an argoFloats object.

i
a character value that specifies the item to be looked up; see “Details”.

j
supplemental index value, ignored unless i is "profile" (see “Details”).

...
ignored.

Details

If i=="metadata" then the metadata slot of x is returned.

If i=="data" then the data slot of x is returned.

If i=="ftpRoot" and the object was created with getIndex(), then then the value of ftpRoot in the metadata slot is returned. (This is used by getProfiles() to construct a vector of URLs to download.)

If i=="destdir" then the destdir item in the metadata slot is returned. This is used by getProfiles() to decide where to save downloaded files, and later by readProfiles(), which works with the output from getProfiles().

If i=="type" then the type item in the metadata slot is returned. This is "index" if the object was created with getIndex(), "profiles" if it was created with getProfiles(), and "argos" if it was created with readProfiles().

If i=="profile" and x was created with readProfiles(), then the return value depends on the value of j. There are 4 sub-cases. (a) If j is not supplied, the return values is a list containing all the profiles in x, each an argo object as created by oce::read.argo() in the oce package. (b) If j is a single integer, then the return value is a single argo object. (c) If j is a vector of integers, then a list of argo objects is returned. (d) If j is the string "count", then the number of profiles is returned.

Otherwise, if x was created with getIndex(), i may be the name of an item in the index item in the data slot, or a string that is made up of enough characters to uniquely identify such an item, e.g. "lon" may be used as a shortcut for "longitude".

Examples

data(index)
# Full remote filename for first two item in index
paste(index[["ftpRoot"]], index[["profile", 1:2]], sep="/")

#> [1] "ftp://ftp.ifremer.fr/ifremer/argo/dac/aoml/1901584/profiles/R1901584_124.nc"
#> [2] "ftp://ftp.ifremer.fr/ifremer/argo/dac/aoml/1901584/profiles/R1901584_125.nc"

# File names and geographical locations of first 5 items in index
index5 <- subset(index, 1:5)
data.frame(file=gsub(".*/", "", index5[["file"]][1]),
           lon=index5[["longitude"]],
           lat=index5[["latitude"]])

#>              file     lon    lat
#> 1 R1901584_124.nc -76.231 27.705
#> 2 R1901584_124.nc -76.024 27.773
#> 3 R1901584_124.nc -76.320 28.064
#> 4 R1901584_124.nc -76.520 28.101
#> 5 R1901584_124.nc -77.108 27.794