Switch [[ and Plot to Focus on Adjusted data

This function returns a copy of its first argument that has been modified by (optionally) replacing 'raw' parameter values, and their flags, with corresponding 'adjusted' values. The purpose is to ensure that future calls to [[,argoFloats-method and plotting methods will focus on the adjusted data instead of the raw data. The procedure hinges on the value of the fallback argument.

useAdjustedNEW(argos, fallback = "NA", debug = 0)

Arguments

argos
an argoFloats object, as read by readProfiles().

fallback
a character value indicating what to do if all the adjusted values for a particular parameter-profile pair are NA. The choices are "NA" and "raw"; see “Details”.

debug
an integer that, if positive, indicates that some debugging information should be printed.

Details

There are two cases, depending on the value of fallback.

If fallback is "NA" (the default), then the adjusted values become the focus, even if they are all NA. Thus, fallback="NA" may be preferred as part of a cautious analysis that focuses only on data of high quality. The downside of this approach is a reduction in coverage, since it eliminates the 'raw' fields that are typical of real-time mode data components (see Section 2.2.5 of Reference 1).

If fallback is "raw", then the data in the adjusted parameter data columns (corresponding to profiles) of each cycle are examined one by one. If the values in a given column are all NA, then the parameter's raw values for that column are not altered. However, if any of the values in a given adjusted column are non-NA, then that entire 'adjusted' column is copied into the corresponding 'raw' column. Since the fallback="raw" scheme combines 'raw' and 'adjusted' data, it increases coverage, at the possible expense of lowering overall data quality.

References

Argo Data Management Team. "Argo User’s Manual V3.4," January 20, 2021. https://archimer.ifremer.fr/doc/00187/29825/

Author

Dan Kelley and Jaimie Harbin