Sample argo file (synthetic realtime data)

This is the netcdf file for cycle 131 of Argo float 2902204, downloaded from ftp://ftp.ifremer.fr/ifremer/argo/dac/incois/2902204/profiles/SR2902204_131.nc on 2020 June 24. As its filename indicates, it holds the "synthetic" version of "realtime" data.

See also

Examples

library(argoFloats)
a <- readProfiles(system.file("extdata", "SR2902204_131.nc", package="argoFloats"))

#> Warning: Of 1 profiles read, 1 has >10% of BBP700 values with QC flag of 4, signalling bad data.
#>     The indices of the bad profiles are as follows.
#>     1

#> Warning: Of 1 profiles read, 1 has >10% of chlorophyllA values with QC flag of 4, signalling bad data.
#>     The indices of the bad profiles are as follows.
#>     1

#> Warning: Of 1 profiles read, 1 has >10% of oxygen values with QC flag of 4, signalling bad data.
#>     The indices of the bad profiles are as follows.
#>     1

#> Warning: Of 1 profiles read, 1 has >10% of pressure values with QC flag of 4, signalling bad data.
#>     The indices of the bad profiles are as follows.
#>     1

summary(a)

#> argoFloats summary
#> ------------------
#> 
#> * type:        argos
#> * contains 1 objects
#> * hint: access list of oce::argo objects with x[["argos"]] or a single one with e.g. x[[1]]
#> * Processing Log
#> 
#>     - 2021-06-30 11:49:03 UTC: `create 'argoFloats' object`
#>     - 2021-06-30 11:49:03 UTC: `readProfiles(profiles = system.file("extdata", "SR2902204_131.nc",     package = "argoFloats"))`

summary(a[[1]])

#> Argo Summary
#> ------------
#> 
#> * Source:              "/Users/runner/work/_temp/Library/argoFloats/extdata/SR2902204_131.nc"
#> * id:                  "2902204"
#> * feature type:        "trajectoryProfile"
#> * Profiles:            0 delayed; 0 adjusted; 0 realtime
#> * Time:                2018-01-23 18:18:35
#> * Data Overview
#> 
#>                                                 Min.       Mean      Max.  Dim. NAs          OriginalName
#>     latitude [°N]                             21.041     21.041    21.041     1   0              LATITUDE
#>     longitude [°E]                             66.67      66.67     66.67     1   0             LONGITUDE
#>     pressure [dbar]                             3.99     253.15     528.2 335x1   0                  PRES
#>     pressureAdjusted [dbar]                     4.03     253.19    528.24 335x1   0         PRES_ADJUSTED
#>     pressureAdjustedError [dbar]                  NA         NA        NA 335x1 335   PRES_ADJUSTED_ERROR
#>     temperature [°C, ITS-90]                  12.826     17.617    24.504 335x1   0                  TEMP
#>     temperatureAdjusted [°C, ITS-90]          12.826     17.617    24.504 335x1   0         TEMP_ADJUSTED
#>     temperatureAdjustedError [°C, ITS-90]         NA         NA        NA 335x1 335   TEMP_ADJUSTED_ERROR
#>     salinity [PSS-78]                         35.582      35.94    36.224 335x1   0                  PSAL
#>     salinityAdjusted [PSS-78]                 35.582      35.94    36.224 335x1   0         PSAL_ADJUSTED
#>     salinityAdjustedError [PSS-78]                NA         NA        NA 335x1 335   PSAL_ADJUSTED_ERROR
#>     oxygen [μmol/kg]                        -0.65686     35.741    200.11 335x1 263                  DOXY
#>     oxygenAdjusted [μmol/kg]                      NA         NA        NA 335x1 335         DOXY_ADJUSTED
#>     oxygenAdjustedError [μmol/kg]                 NA         NA        NA 335x1 335   DOXY_ADJUSTED_ERROR
#>     chlorophyllA                              0.0511    0.35557    2.0878 335x1 263                  CHLA
#>     chlorophyllAAdjusted                          NA         NA        NA 335x1 335         CHLA_ADJUSTED
#>     chlorophyllAAdjustedError                     NA         NA        NA 335x1 335   CHLA_ADJUSTED_ERROR
#>     BBP700                                0.00032344 0.00070775 0.0018912 335x1 263                BBP700
#>     BBP700Adjusted                                NA         NA        NA 335x1 335       BBP700_ADJUSTED
#>     BBP700AdjustedError                           NA         NA        NA 335x1 335 BBP700_ADJUSTED_ERROR
#> 
#> * Data-quality Flag Scheme
#> 
#>     name    "argo"
#>     mapping list(not_assessed=0, passed_all_tests=1, probably_good=2, probably_bad=3, bad=4, changed=5, not_used_6=6, not_used_7=7, estimated=8, missing=9)
#>     default c(0, 3, 4, 6, 7, 9)
#> 
#> * Data-quality Flags
#> 
#>     pressure:             "1" 263, "3" 72
#>     pressureAdjusted:     "1" 263, "3" 1, "8" 71
#>     temperature:          "1" 263, "3" 1, "8" 71
#>     temperatureAdjusted:  "1" 263, "3" 1, "8" 71
#>     salinity:             "1" 263, "3" 1, "8" 71
#>     salinityAdjusted:     "1" 263, "3" 1, "8" 71
#>     oxygen:               "1" 37, "3" 1, "4" 34
#>     oxygenAdjusted:       NA 335 
#>     chlorophyllA:         "0" 72
#>     chlorophyllAAdjusted: NA 335 
#>     BBP700:               "0" 72
#>     BBP700Adjusted:       NA 335 
#> 
#> * Processing Log
#> 
#>     - 2021-06-30 11:49:03 UTC: `create 'argo' object`
#>     - 2021-06-30 11:49:03 UTC: `initializeFlagScheme(object, name="argo", mapping=list(not_assessed=0,passed_all_tests=1,probably_good=2,probably_bad=3,bad=4,changed=5,not_used_6=6,not_used_7=7,estimated=8,missing=9)), default=c(0,3,4,9))`
#>     - 2021-06-30 11:49:03 UTC: `read.argo(file="/Users/runner/work/_temp/Library/argoFloats/extdata/SR2902204_131.nc")`
#>     - 2021-06-30 11:49:03 UTC: `override existing flagScheme to be mapping=list(not_assessed=0, passed_all_tests=1, probably_good=2, probably_bad=3, bad=4, changed=5, not_used_6=6, not_used_7=7, estimated=8, missing=9)),  default=c(0, 3, 4, 9)`