Get Default Values

By default, files downloaded from the Argo mirror are placed in ~/data/argo. You can configure where these files are stored by setting options(argoFloats.destdir="path/to/argo"). To persist this value between sessions you can place this value in your .Rprofile. Similarly, the default value for server in getIndex() is "ifremer-https", which can be configured by setting options(argoFloats.server = "https://...").

argoDefaultDestdir()

argoDefaultServer()

argoDefaultIndexAge()

argoDefaultProfileAge()

hasArgoTestCache()

Value

A character value containing a path to the default Argo data directory.

Examples

argoDefaultDestdir()

#> [1] "/Users/runner/work/_temp/argo-cache-dir"

argoDefaultServer()

#> [1] "ifremer-https"

argoDefaultIndexAge()

#> [1] Inf

argoDefaultProfileAge()

#> [1] Inf