Interactive app for viewing Argo float positions

The GUI permits specifying a spatial-temporal region of interest, a set of float types to show, etc. The interface ought to be reasonably straightforward, especially for those who take a moment to click on the Help button and to read the popup window that it creates.

mapApp(
  age = argoDefaultIndexAge(),
  destdir = argoDefaultDestdir(),
  server = "auto",
  debug = 0
)

Arguments

age
numeric value indicating how old a downloaded file must be (in days), for it to be considered out-of-date. The default, argoDefaultIndexAge(), limits downloads to once per day, as a way to avoid slowing down a workflow with a download that might take a minute or so. Note that setting age=0 will force a new download, regardless of the age of the local file.

destdir
character value indicating the directory in which to store downloaded files. The default value is to compute this using argoDefaultDestdir(), which returns ~/data/argo by default, although it also provides ways to set other values using options(). Set destdir=NULL if destfile is a filename with full path information. File clutter is reduced by creating a top-level directory called data, with subdirectories for various file types; see “Examples”.

server
character value, or vector of character values, indicating the name of servers that supply argo data acquired with getIndex(). The default value, "auto", indicates to try a sequence of servers.

debug
integer value that controls how much information mapApp() prints to the console as it works. The default value of 0 leads to a fairly limited amount of printing, while higher values lead to more information. This information can be helpful in diagnosing problems or bottlenecks.

Details

This app will use getIndex() to download index files from the Argo server the first time it runs, and this make take up to a minute or so. Then it will combine information from the core-Argo and bgc-Argo index tables, cross-indexing so it can determine the Argo type for each profile (or cycle).

The hi-res button will only affect the coastline, not the topography, unless there is a local file named topoWorldFine.rda that contains an alternative topographic information. Here is how to create such a file:

library(oce)
topoFile <- download.topo(west=-180, east=180,
                          south=-90, north=90,
                          resolution=10,
                          format="netcdf", destdir=".")
topoWorldFine <- read.topo(topoFile)
save(topoWorldFine, file="topoWorldFine.rda")
unlink(topoFile) # clean up

Author

Dan Kelley

Examples

if (FALSE) {
library(argoFloats)
mapApp()}