Read argo profiles from local files

This works with either a list of local (netCDF) files, or a argoFloats object of type "profiles", as created by getProfiles(). By default, warnings are issued about any profiles in which 50 percent or more of the measurements are flagged with a quality-control code of 4 (which designates bad data), and these values are set to NA; use the silent and handleFlags arguments to control this behaviour.

readProfiles(profiles, handleFlags = TRUE, silent = FALSE, debug = 0)

Arguments

profiles
either a character vector holding the names of local files to read, or (better) an argoFloats object created by getProfiles().

handleFlags
a logical value (with default TRUE) that indicates whether to call oce::handleFlags() on the individual argo files that are to be read. This cleans up some common errors that are identified in the quality-control analysis that is usually done before data are place on the argo servers.

silent
logical value (with FALSE as the default value), indicating whether to silence messages and warnings. The value of silent is ignored if debug is TRUE.

debug
an integer specifying the level of debugging. If this is zero, the work proceeds silently. If it is 1, a small amount of debugging information is printed. Note that debug-1 is passed to oce::read.argo(), which actually reads the file, and so it will print messages if debug exceeds 1.

Value

readProfiles returns an argoFloats object with type="argos", in which the data slot contains a list named argos that holds objects that are created by oce::read.argo().

Examples

# Download and plot some profiles.
if (FALSE) {
library(argoFloats)
data(index)
index1 <- subset(index, 1)
profiles <- getProfiles(index1)
argosWithNA<- readProfiles(profiles, handleFlags=TRUE)
argosWithoutNA <- readProfiles(profiles, handleFlags=FALSE)
par(mfrow=c(1, 2))
file <- gsub(".*/", "",  profiles[[1]])
aWithNA <- argosWithNA[[1]]
plotTS(aWithNA, eos="unesco", type="o")
mtext(paste(file, "\n handling flags"), cex=0.7*par("cex"))
aWithoutNA <- argosWithoutNA[[1]]
plotTS(aWithoutNA, eos="unesco", type="o")
mtext(paste(file, "\n ignoring flags"), cex=0.7*par("cex"))
}